rename smoke test to web_ui and run in CI
"smoke" is fairly well known as a first test to run (to check if it
smokes), but this one is focussed on just the Web UI not testing
the overall project.

Also:
+ add requirements in Pipfile
+ handle starting server in the Python code, so we can reliably
  stop it and not also stop a voila instance the developer happens
  to be running.
