--- start.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-./stop.sh
-voila --enable_nbextensions=True ./VoilaDrawdown.ipynb --no-browser &
