--- stop.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-pkill voila
