+++ tox.ini
@@ -20,13 +20,11 @@ commands =
 # runs on github workflow
 [testenv:ci]
 commands =
-    python -m pytest --cov=. --cov-report term --cov-report=xml --ignore=tests/test_smoke.py
+    python -m pytest --cov=. --cov-report term --cov-report=xml
 
-[testenv:smoke]
+[testenv:web_ui]
 commands =
-    ./start.sh
-    python -m pytest tests/test_smoke.py
-    ./stop.sh
+    python -m pytest tests/test_web_ui.py
 
 [testenv:health_update]
 commands =
